projects/AI Travel Planner Agent/requirements.txt
langchain
langchain-groq
langgraph
fastapi
uvicorn
pydantic-settings
